rewrite and publicize minecraft server commands
These commands have existed for a while but were kept private due to them only being used in a private discord server and not the structure not handling different servers (if someone set an ip for their server, anyone who used the status command would see the result for that ip). Now that I have reset my computer and lost local changes I decided to rewrite it (better this time) and publish it.